feat(color): Use generic YAML parser to save/load Theme yaml files (#3346)
* Change ThemeFile class to use generic YAML parser instead of hard wired code.
Fix handling of empty input lines in YAML parser.
Increase size of YAML parser variables to handle theme 'info' property length.

* Update comment.

* Fix error.

* Proper fix for blank lines in YAML file.

* Fix YAML blank line handling (attempt 3).

* Ensure theme YAML file is backward compatible.

* Store selected theme in radio settings instead of selectedtheme.txt file.
On start, if radio settings theme is empty, try and load from selectedtheme.txt file.
In Companion, save/load the selected theme in the radio settings YAML file; but no UI to select theme.

* Update selected theme name in radio.yml if the theme name was edited.

* Only save 'selectedTheme' in Companion on color LCD radios.

+++ radio/src/storage/yaml/yaml_tree_walker.h
@@ -156,7 +156,7 @@ public:
 
     bool isElmtEmpty(uint8_t* data);
 
-    void setAttrValue(char* buf, uint8_t len);
+    void setAttrValue(char* buf, uint16_t len);
 
     bool generate(yaml_writer_func wf, void* opaque);
